src/backend/requirements.txt
django
django-rest-framework
psycopg2